11 - Refactor script to be in utils
