+++ Machine Learning for Trading/6 Histograms and Scatter Plots/plot-histograms.py
@@ -1,39 +1,8 @@
 """Plot a histogram"""
 
-import os
 import pandas as pd
-import matplotlib.pyplot as plt
 
-# from util import get_data, plot_data
-
-
-def symbol_to_path(symbol, base_dir="data"):
-    """Return CSV file path given ticker symbol."""
-    return os.path.join(base_dir, "{}.csv".format(str(symbol)))
-
-
-def get_data(symbols, dates):
-    """Read stock data (adjusted close) for given symbols from CSV files."""
-    print("Getting data...")
-    df_final = pd.DataFrame(index=dates)
-
-    for symbol in symbols:
-        file_path = symbol_to_path(symbol)
-        df_temp = pd.read_csv(file_path, parse_dates=True, index_col="Date",
-                              usecols=["Date", "Adj Close"], na_values=["nan"])
-        df_temp = df_temp.rename(columns={"Adj Close": symbol})
-        df_final = df_final.join(df_temp)
-
-    return df_final
-
-
-def plot_data(data, title,
-              ylabel, xlabel):
-    """Plot stock data with appropriate axis labels."""
-    ax = data.plot(title=title, fontsize=12)
-    ax.set_xlabel(xlabel)
-    ax.set_ylabel(ylabel)
-    plt.show()
+from utils.stock import get_data, plot_data
 
 
 def compute_daily_returns(df):
