+++ Machine Learning for Trading/6 Histograms and Scatter Plots/utils/stock.py
@@ -1,0 +1,31 @@
+# stock.py
+import os
+import pandas as pd
+import matplotlib.pyplot as plt
+
+
+def symbol_to_path(symbol, base_dir="data"):
+    """Return CSV file path given ticker symbol."""
+    return os.path.join(base_dir, "{}.csv".format(str(symbol)))
+
+
+def get_data(symbols, dates):
+    """Read stock data (adjusted close) for given symbols from CSV files."""
+    df_final = pd.DataFrame(index=dates)
+
+    for symbol in symbols:
+        file_path = symbol_to_path(symbol)
+        df_temp = pd.read_csv(file_path, parse_dates=True, index_col="Date",
+                              usecols=["Date", "Adj Close"], na_values=["nan"])
+        df_temp = df_temp.rename(columns={"Adj Close": symbol})
+        df_final = df_final.join(df_temp)
+
+    return df_final
+
+
+def plot_data(data, title,
+              ylabel, xlabel):
+    ax = data.plot(title=title, fontsize=12)
+    ax.set_xlabel(xlabel)
+    ax.set_ylabel(ylabel)
+    plt.show()
